+++ UnitTests/BraintreeTestShared/MockAPIClient.swift
@@ -27,6 +27,7 @@ public class MockAPIClient: BTAPIClient {
     public var postedDidPayPalServerAttemptAppSwitch: Bool? = nil
     public var postedErrorDescription: String? = nil
     public var postedContextType: String? = nil
+    public var postedFundingSource: String? = nil
     
     @objc public var cannedConfigurationResponseBody : BTJSON? = nil
     @objc public var cannedConfigurationResponseError : NSError? = nil
@@ -50,6 +51,21 @@ public class MockAPIClient: BTAPIClient {
         completionBlock(cannedResponseBody, cannedHTTPURLResponse, cannedResponseError)
     }
     
+    public override func get(
+        _ path: String,
+        parameters: Encodable?,
+        httpType: BTAPIClientHTTPService
+    ) async throws -> (BTJSON?, HTTPURLResponse?) {
+        lastGETPath = path
+        lastGETParameters = try? parameters?.toDictionary()
+        lastGETAPIClientHTTPType = httpType
+        
+        if let error = cannedResponseError {
+            throw error
+        }
+        return (cannedResponseBody, cannedHTTPURLResponse)
+    }
+    
     public override func post(_ path: String, parameters: Encodable? = nil, headers: [String: String]? = nil, httpType: BTAPIClientHTTPService, completion completionBlock: ((BTJSON?, HTTPURLResponse?, Error?) -> Void)? = nil) {
         lastPOSTPath = path
         lastPOSTParameters = try? parameters?.toDictionary()
@@ -61,6 +77,24 @@ public class MockAPIClient: BTAPIClient {
         }
         completionBlock(cannedResponseBody, cannedHTTPURLResponse, cannedResponseError)
     }
+
+    public override func post(
+        _ path: String,
+        parameters: Encodable? = nil,
+        headers: [String: String]? = nil,
+        httpType: BTAPIClientHTTPService = .gateway
+    ) async throws -> (BTJSON?, HTTPURLResponse?) {
+        lastPOSTPath = path
+        lastPOSTParameters = try? parameters?.toDictionary()
+        lastPOSTAPIClientHTTPType = httpType
+        lastPOSTAdditionalHeaders = headers
+        
+        if let error = cannedResponseError {
+            throw error
+        }
+        
+        return (cannedResponseBody, cannedHTTPURLResponse)
+    }
     
     public override func fetchOrReturnRemoteConfiguration(_ completionBlock: @escaping (BTConfiguration?, Error?) -> Void) {
         guard let responseBody = cannedConfigurationResponseBody else {
@@ -70,7 +104,7 @@ public class MockAPIClient: BTAPIClient {
         completionBlock(BTConfiguration(json: responseBody), cannedConfigurationResponseError)
     }
     
-    public override func fetchConfiguration() async throws -> BTConfiguration {
+    public override func fetchOrReturnRemoteConfiguration() async throws -> BTConfiguration {
         guard let responseBody = cannedConfigurationResponseBody else {
             throw cannedConfigurationResponseError ?? NSError(domain: "com.example.error", code: -1, userInfo: nil)
         }
@@ -89,10 +123,11 @@ public class MockAPIClient: BTAPIClient {
         didEnablePayPalAppSwitch: Bool? = nil,
         didPayPalServerAttemptAppSwitch: Bool? = nil,
         errorDescription: String? = nil,
-        merchantExperiment experiment: String? = nil,
+        fundingSource: String? = nil,
         isConfigFromCache: Bool? = nil,
         isVaultRequest: Bool? = nil,
         linkType: LinkType? = nil,
+        merchantExperiment experiment: String? = nil,
         pageType: String? = nil,
         shopperSessionID: String? = nil
     ) {
@@ -110,7 +145,8 @@ public class MockAPIClient: BTAPIClient {
         postedDidPayPalServerAttemptAppSwitch = didPayPalServerAttemptAppSwitch
         postedErrorDescription = errorDescription
         postedContextType = contextType
-
+        postedFundingSource = fundingSource
+        
         postedAnalyticsEvents.append(eventName)
     }